feat: data in project details & deleted unnecessary
