--- src/components/pages/ProjectsPrivatePage/ProjectsPrivatePage.styles.ts
@@ -1,7 +1,0 @@
-import { styled, Container } from '@mui/material';
-
-export const ContainerWrapper = styled(Container)(() => ({
-  display: 'flex',
-  flexDirection: 'column',
-  rowGap: 12,
-})) as typeof Container;
