--- src/components/pages/ProjectsPrivatePage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProjectsPrivatePage';
